--- a/rcl/test/rcl/test_wait.cpp
+++ b/rcl/test/rcl/test_wait.cpp
@@ -144,13 +144,12 @@ TEST(CLASSNAME(WaitSetTestFixture, RMW_IMPLEMENTATION), zero_timeout) {
 
 // Check rcl_wait can be called in many threads, each with unique wait sets and resources.
 TEST(CLASSNAME(WaitSetTestFixture, RMW_IMPLEMENTATION), multi_wait_set_threaded) {
-  auto start = std::chrono::steady_clock::now();
   rcl_ret_t ret;
   const size_t number_of_threads = 20;  // concurrent waits
   const size_t count_target = 10;  // number of times each wait should wake up before being "done"
   const size_t retry_limit = 100;  // number of times to retry when a timeout occurs waiting
-  const uint64_t wait_period = RCL_MS_TO_NS(500);  // timeout passed to rcl_wait each try
-  const std::chrono::milliseconds trigger_period(1);  // period between each round of triggers
+  const uint64_t wait_period = RCL_MS_TO_NS(100);  // timeout passed to rcl_wait each try
+  const std::chrono::milliseconds trigger_period(2);  // period between each round of triggers
   struct TestSet
   {
     std::atomic<size_t> wake_count;
@@ -181,7 +180,6 @@ TEST(CLASSNAME(WaitSetTestFixture, RMW_IMPLEMENTATION), multi_wait_set_threaded)
             // printf("%s\n", ss.str().c_str());
           }
           ret = rcl_wait(&test_set.wait_set, wait_period);
-          ASSERT_NE(ret, RCL_RET_ERROR);
           if (ret != RCL_RET_TIMEOUT) {
             {
               std::stringstream ss;
@@ -198,6 +196,7 @@ TEST(CLASSNAME(WaitSetTestFixture, RMW_IMPLEMENTATION), multi_wait_set_threaded)
             // no need to wait again
             break;
           } else {
+            ASSERT_EQ(ret, RCL_RET_OK);
             std::stringstream ss;
             ss << "[thread " << test_set.thread_id << "] Timeout (try #" << wake_try_count << ")";
             // printf("%s\n", ss.str().c_str());
@@ -297,7 +296,6 @@ TEST(CLASSNAME(WaitSetTestFixture, RMW_IMPLEMENTATION), multi_wait_set_threaded)
   //   loop_count);
 
   auto end = std::chrono::steady_clock::now();
-  printf("Test finished in %zu nanoseconds\n", (end - start).count());
 }
 
 // Check the interaction of a guard condition and a negative timeout by