We previously discussed problems with `std::this_thread::sleep` waking up spuriously on Windows. In this PR I add a spinning wait as a utility in the rcl tests, use it in the trigger thread, use `std::this_thread::yield` before going to sleep to provide hints to the Windows thread scheduler for when to context switch, and increase the trigger period. 

CI jobs with arguments `--end-with rcl --ctest-args --repeat-until-fail 20 -R test_wait__rmw_opensplice_cpp --`

Windows, Build type=None
[![Build Status](http://ci.ros2.org/job/ci_windows/1412/badge/icon)](http://ci.ros2.org/job/ci_windows/1412/)

Windows, Build type=Release
[![Build Status](http://ci.ros2.org/job/ci_windows/1413/badge/icon)](http://ci.ros2.org/job/ci_windows/1413/)

(problem is not present on Debug job)