@jacquelinekay 2016-06-06
ok, through instrumenting the call to sleep_for I convinced myself that a custom sleep_for is not needed, so I will remove the sleep_for.hpp header.

In order to convince myself one way or the other about `yield`, I collected a small amount of data (16 runs). I timed how long the entire test took when "wait_period" was equal to 5ms and "trigger_period" was equal to 10ms. @wjwwood's hypothesis was that calls to `yield` were introducing a delay that happened to make the test pass and therefore aren't needed.

With `std::this_thread::yield` called before `sleep_for` and `rcl_wait`, the tests took an average of 466 ms, and the test succeeded 100% of the time.
Without `std::this_thread::yield`, the tests took an average of 906 ms. I had to run the test 30 times to get 16 successes and 14 failures (I didn't count timing information for the failures).

16 samples each might not be statistically significant but to me this suggests that the Windows thread scheduler doesn't necessarily know to context switch without `yield`. It also isn't specified in the C++ standard that `sleep_for` has to force a context switch so I see no specification for why that would have to be (other than common sense).

There certainly are more precise ways to measure this, but the system calls for counting context switches in Windows look kind of awful and I didn't delve too deeply into them. I can make a standalone test with only the standard library calls if others think I need more evidence.