@jacquelinekay 2016-06-06
I've decided on a fix that William and I can probably agree on:

I reverted my other changes and changed `wait_period` to a very large number (100ms, compared to the trigger_period of 2ms). I think this test was flaky before because after rcl_wait timed out the test_set threads would continuously try waiting again but the trigger thread would not get rescheduled, leading to the trigger condition never happening. By significantly increasing the wait time, the trigger thread will continuously trigger and the test_set threads will wait and wake up when they received a guard condition. I think this is also fairly consistent with a common use case for `rcl_wait`, which is to call it from `spin`, which will pass rcl_wait a "block indefinitely until woken up" argument (-1). Does that make sense?