@jacquelinekay 2016-06-06
[![Build Status](http://ci.ros2.org/job/ci_windows/1418/badge/icon)](http://ci.ros2.org/job/ci_windows/1418/) (with arguments `--end-with rcl --ctest-args --repeat-until-fail 20 -R test_wait__rmw_opensplice_cpp --` to check if the test is still flaky)