@wjwwood 2016-06-07
+1 It makes sense to me! Thanks for iterating on this with me @jacquelinekay.